refactor: Controller 기능 세분화
