+++ src/main/java/roomescape/controller/PageController.java
@@ -1,0 +1,19 @@
+package roomescape.controller;
+
+import org.springframework.stereotype.Controller;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.RequestMapping;
+
+@Controller
+@RequestMapping("/admin")
+public class PageController {
+    @GetMapping()
+    public String welcomePage() {
+        return "/admin/index";
+    }
+
+    @GetMapping("/reservation")
+    public String reservationPage() {
+        return "/admin/reservation-legacy";
+    }
+}
